separate sidebar to new component
